% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_languages.R
\name{available_languages}
\alias{available_languages}
\title{List available programming languages}
\usage{
available_languages()
}
\description{
List programming languages for which winners can be retrieved.
}
