% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r_package_authors.R
\name{get_r_package_authors}
\alias{get_r_package_authors}
\title{Retrieve Author Names}
\usage{
get_r_package_authors(pkg.name)
}
\arguments{
\item{pkg.name}{name of the R package for which author names shall be retrieved.}
}
\description{
R package author name retrieval
}
\examples{
get_r_package_authors("MASS")
}
