% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_winners.R
\name{list_winners}
\alias{list_winners}
\title{Winner Retrieval}
\usage{
list_winners(language = "R")
}
\arguments{
\item{language}{programming language for which winners shall be retrieved.}
}
\description{
Retrieval of winner names.
}
